+++ flask_unchained/bundles/controller/extensions.py
@@ -1,0 +1,15 @@
+from flask_wtf import CSRFProtect
+
+
+csrf = CSRFProtect()
+
+
+EXTENSIONS = {
+    'csrf': csrf,
+}
+
+
+__all__ = [
+    'csrf',
+    'CSRFProtect',
+]