+++ flask_unchained/bundles/security/extensions/__init__.py
@@ -1,21 +1,15 @@
-from flask_wtf import CSRFProtect
-
 from .security import Security
 
 
-csrf = CSRFProtect()
 security = Security()
 
 
 EXTENSIONS = {
-    'csrf': csrf,
     'security': (security, ['csrf', 'db'])
 }
 
 
 __all__ = [
-    'csrf',
-    'CSRFProtect',
     'security',
     'Security',
 ]